Remove RatisServerConfig.
