+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/ratis/OzoneManagerRatisServerConfig.java
@@ -20,7 +20,6 @@ package org.apache.hadoop.ozone.om.ratis;
 import org.apache.hadoop.hdds.conf.Config;
 import org.apache.hadoop.hdds.conf.ConfigGroup;
 import org.apache.hadoop.hdds.conf.ConfigType;
-import org.apache.hadoop.hdds.conf.RatisServerConfig;
 import org.apache.hadoop.ozone.om.OMConfigKeys;
 import org.apache.ratis.server.RaftServerConfigKeys;
 
@@ -28,6 +27,7 @@ import java.time.Duration;
 
 import static org.apache.hadoop.hdds.conf.ConfigTag.OM;
 import static org.apache.hadoop.hdds.conf.ConfigTag.OZONE;
+import static org.apache.hadoop.hdds.conf.ConfigTag.PERFORMANCE;
 import static org.apache.hadoop.hdds.conf.ConfigTag.RATIS;
 
 /**
@@ -35,7 +35,23 @@ import static org.apache.hadoop.hdds.conf.ConfigTag.RATIS;
  */
 @ConfigGroup(prefix = OMConfigKeys.OZONE_OM_HA_PREFIX + "."
     + RaftServerConfigKeys.PREFIX)
-public class OzoneManagerRatisServerConfig extends RatisServerConfig {
+public class OzoneManagerRatisServerConfig {
+  /** @see RaftServerConfigKeys.Log.Appender#WAIT_TIME_MIN_KEY */
+  @Config(key = "log.appender.wait-time.min",
+      defaultValue = "1ms",
+      type = ConfigType.TIME,
+      tags = {OZONE, OM, RATIS, PERFORMANCE},
+      description = "Minimum wait time between two appendEntries calls."
+  )
+  private long logAppenderWaitTimeMin;
+
+  public long getLogAppenderWaitTimeMin() {
+    return logAppenderWaitTimeMin;
+  }
+
+  public void setLogAppenderWaitTimeMin(long logAppenderWaitTimeMin) {
+    this.logAppenderWaitTimeMin = logAppenderWaitTimeMin;
+  }
 
   @Config(key = "retrycache.expirytime",
       defaultValue = "300s",
